доделал проверку на выход за пределы массива
