+++ Java3GeekBrains/src/Lesson1/ChangeElements/Man.java
